订单模块第一版修改

+++ begin-order/src/main/java/com/example/begin/order/service/impl/OrderServiceImpl.java
@@ -111,7 +111,7 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public Result queryOrderById(int id) {
+    public Result queryOrderById(long id) {
         return null;
     }
